+++ include/Addon/AddonUpdater.php
@@ -52,7 +52,13 @@ class AddonUpdater {
 	 */
 	public function setup_updater() {
 		$email_log = email_log();
-		$license_key = $email_log->get_licenser()->get_addon_license_key( $this->addon_name );
+		$licenser = $email_log->get_licenser();
+
+		if ( is_null( $licenser ) ) {
+			return;
+		}
+
+		$license_key = $licenser->get_addon_license_key( $this->addon_name );
 
 		$updater = new EDDUpdater( $email_log->get_store_url(), $this->addon_file, array(
 				'version'   => $this->addon_version,
@@ -62,6 +68,6 @@ class AddonUpdater {
 			)
 		);
 
-		$email_log->get_licenser()->add_updater( $updater );
+		$licenser->add_updater( $updater );
 	}
 }