+++ include/Core/EmailLogger.php
@@ -39,16 +39,20 @@ class EmailLogger implements Loadie {
 		 */
 		$mail_info = apply_filters( 'el_wp_mail_log', $mail_info );
 
-		$headers = '';
-		if ( isset( $mail_info['headers'] ) ) {
-			$headers = is_array( $mail_info['headers'] ) ? implode( "\n", $mail_info['headers'] ) : $mail_info['headers'];
-		}
+		// Sometimes the array passed to the `wp_mail` filter may not contain all the required keys.
+		// See https://wordpress.org/support/topic/illegal-string-offset-attachments/
+		$mail_info = wp_parse_args( $mail_info, array(
+			'attachments' => array(),
+			'to'          => '',
+			'subject'     => '',
+			'headers'     => '',
+		) );
 
 		$data = array(
 			'attachments' => ( count( $mail_info['attachments'] ) > 0 ) ? 'true' : 'false',
 			'to_email'    => is_array( $mail_info['to'] ) ? implode( ',', $mail_info['to'] ) : $mail_info['to'],
 			'subject'     => $mail_info['subject'],
-			'headers'     => $headers,
+			'headers'     => is_array( $mail_info['headers'] ) ? implode( "\n", $mail_info['headers'] ) : $mail_info['headers'],
 			'sent_date'   => current_time( 'mysql' ),
 		);