+++ include/Core/UI/Setting/Setting.php
@@ -21,11 +21,12 @@ abstract class Setting {
 	 */
 	public function __construct() {
 		$this->section = new SettingSection();
+
+		$this->initialize();
+
 		$this->section->fields = $this->get_fields();
 		$this->section->callback = array( $this, 'render' );
 		$this->section->sanitize_callback = array( $this, 'sanitize' );
-
-		$this->initialize();
 	}
 
 	/**
@@ -50,11 +51,14 @@ abstract class Setting {
 
 	/**
 	 * Get the value stored in the option.
+	 * If no values are found then the default values are returned.
 	 *
-	 * @return mixed Stored value.
+	 * @return array Stored value.
 	 */
 	public function get_value() {
-		return get_option( $this->section->option_name );
+		$value = get_option( $this->section->option_name );
+
+		return wp_parse_args( $value, $this->section->default_value );
 	}
 
 	/**
@@ -69,14 +73,18 @@ abstract class Setting {
 	 *
 	 * @return SettingField[] List of fields for the Setting.
 	 */
-	abstract protected function get_fields();
+	protected function get_fields() {
+		return $this->build_fields();
+	}
 
 	/**
-	 * Render the Settings.
+	 * Render the Settings section.
 	 *
-	 * @return void
+	 * By default it does nothing.
 	 */
-	abstract public function render();
+	public function render() {
+		return;
+	}
 
 	/**
 	 * Sanitize the option values.
@@ -85,5 +93,47 @@ abstract class Setting {
 	 *
 	 * @return mixed Sanitized values.
 	 */
-	abstract public function sanitize( $values );
+	public function sanitize( $values ) {
+		if ( ! is_array( $values ) ) {
+			return array();
+		}
+
+		$values = wp_parse_args( $values, $this->section->default_value );
+		$sanitized_values = array();
+
+		foreach ( $this->section->field_labels as $field_id => $label ) {
+			$callback = array( $this, 'sanitize_' . $field_id );
+
+			if ( is_callable( $callback ) ) {
+				$sanitized_values[ $field_id ] = call_user_func( $callback, $values[ $field_id ] );
+			} else {
+				$sanitized_values[ $field_id ] = $values[ $field_id ];
+			}
+		}
+
+		return $sanitized_values;
+	}
+
+	/**
+	 * Build SettingField objects from field id and labels.
+	 *
+	 * @since 2.1.0
+	 *
+	 * @return \EmailLog\Core\UI\Setting\SettingField[] Built SettingFields.
+	 */
+	protected function build_fields() {
+		$fields = array();
+
+		foreach ( $this->section->field_labels as $field_id => $label ) {
+			$field           = new SettingField();
+			$field->id       = $field_id;
+			$field->title    = $label;
+			$field->args     = array( 'id' => $field_id );
+			$field->callback = array( $this, 'render_' . $field_id . '_settings' );
+
+			$fields[] = $field;
+		}
+
+		return $fields;
+	}
 }