+++ include/Util/helper.php
@@ -40,10 +40,8 @@ function sanitize_email_with_name( $string ) {
 
 	$bracket_pos = strpos( $string, '<' );
 	if ( false !== $bracket_pos ) {
-		// Text before the bracketed email is the name.
 		if ( $bracket_pos > 0 ) {
 			$name = substr( $string, 0, $bracket_pos );
-			$name = str_replace( '"', '', $name );
 			$name = trim( $name );
 
 			$email = substr( $string, $bracket_pos + 1 );
@@ -63,7 +61,7 @@ function sanitize_email_with_name( $string ) {
  *
  * @since 2.0.0
  *
- * @return array List of Columns to export.
+ * @return string[] List of Columns to export.
  */
 function get_log_columns_to_export() {
 
@@ -72,4 +70,43 @@ function get_log_columns_to_export() {
 	}
 
 	return array( 'id', 'sent_date', 'to_email', 'subject' );
+}
+
+/**
+ * Is it an admin request and not an ajax request.
+ *
+ * @since 2.1
+ *
+ * @return bool True if admin non ajax request, False otherwise.
+ */
+function is_admin_non_ajax_request() {
+	if ( function_exists( 'wp_doing_ajax' ) && wp_doing_ajax() ) {
+		return false;
+	}
+
+	if ( defined( 'DOING_AJAX' ) && DOING_AJAX ) {
+		return false;
+	}
+
+	return is_admin();
+}
+
+/**
+ * Checks the Checkbox when values are present in a given array.
+ *
+ * Use this function in Checkbox fields.
+ *
+ * @since 2.1.0
+ *
+ * @param array $values   List of all possible values.
+ * @param string $current The current value to be checked.
+ */
+function checked_array( $values, $current ) {
+	if ( ! is_array( $values ) ) {
+		return;
+	}
+
+	if ( in_array( $current, $values ) ) {
+		echo "checked='checked'";
+	}
 }