+++ load-email-log.php
@@ -38,18 +38,26 @@ function load_email_log( $plugin_file ) {
 
 	$email_log = new \EmailLog\Core\EmailLog( $plugin_file, $loader, new \EmailLog\Core\DB\TableManager() );
 
-	$email_log->set_licenser( new \EmailLog\Addon\License\Licenser() );
+	if ( \EmailLog\Util\is_admin_non_ajax_request() ) {
+		// Loading licenser in frontend or ajax request is resulting in huge performance issues.
+		$email_log->set_licenser( new \EmailLog\Addon\License\Licenser() );
+
+		$email_log->add_loadie( new \EmailLog\Addon\DependencyEnforcer() );
+		$email_log->add_loadie( new \EmailLog\Core\Request\OverridePluginAPI() );
+	}
 
 	$email_log->add_loadie( new \EmailLog\Core\EmailLogger() );
 	$email_log->add_loadie( new \EmailLog\Core\UI\UILoader() );
-	$email_log->add_loadie( new \EmailLog\Addon\DependencyEnforcer() );
 
 	$email_log->add_loadie( new \EmailLog\Core\Request\NonceChecker() );
 	$email_log->add_loadie( new \EmailLog\Core\Request\LogListAction() );
-	$email_log->add_loadie( new \EmailLog\Core\Request\OverridePluginAPI() );
+
+	$capability_giver = new \EmailLog\Core\AdminCapabilityGiver();
+	$email_log->add_loadie( $capability_giver );
 
 	// `register_activation_hook` can't be called from inside any hook.
 	register_activation_hook( $plugin_file, array( $email_log->table_manager, 'on_activate' ) );
+	register_activation_hook( $plugin_file, array( $capability_giver, 'add_cap_to_admin' ) );
 
 	// Ideally the plugin should be loaded in a later event like `init` or `wp_loaded`.
 	// But some plugins like EDD are sending emails in `init` event itself,