+++ uninstall.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Uninstall page for Email Log Plugin to clean up db.
+ * Uninstall page for Email Log Plugin to clean up all plugin data.
  *
  * This file is named uninstall.php since WordPress requires that name.
  */
@@ -17,31 +17,55 @@ if ( is_multisite() ) {
 
 	foreach ( $sites as $site ) {
 		switch_to_blog( $site['blog_id'] );
-		email_log_delete_table();
+		email_log_delete_db_data();
 		restore_current_blog();
 	}
 } else {
-	email_log_delete_table();
+	email_log_delete_db_data();
 }
 
 /**
- * Delete email log table and db option
+ * Delete all email log data from db.
+ *
+ * The data include email log table, options, capability and add-on license data.
  *
  * @since 1.7
  *
  * @global object $wpdb
  */
-function email_log_delete_table() {
+function email_log_delete_db_data() {
 	global $wpdb;
+
+	$remove_data_on_uninstall = false;
+
+	$option = get_option( 'email-log-core' );
+	if ( is_array( $option ) && array_key_exists( 'remove_on_uninstall', $option ) &&
+	     'true' === strtolower( $option['remove_on_uninstall'] ) ) {
+
+		$remove_data_on_uninstall = true;
+	}
 
 	// This is hardcoded on purpose, since the entire plugin is not loaded during uninstall.
 	$table_name = $wpdb->prefix . 'email_log';
 
-	if ( $wpdb->get_var( "SHOW TABLES LIKE '{$table_name}'" ) == $table_name ) {
-		// If table is present, drop it
-		$wpdb->query( "DROP TABLE $table_name" );
-	}
+	if ( $remove_data_on_uninstall ) {
+		if ( $wpdb->get_var( "SHOW TABLES LIKE '{$table_name}'" ) == $table_name ) {
+			$wpdb->query( "DROP TABLE $table_name" );
+		}
 
-	// Delete the option
-	delete_option( 'email-log-db' );
+		delete_option( 'email-log-db' );
+		delete_option( 'email-log-core' );
+
+		$roles = get_editable_roles();
+		foreach ( $roles as $role_name => $role_obj ) {
+			$role = get_role( $role_name );
+
+			if ( ! is_null( $role ) ) {
+				$role->remove_cap( 'manage_email_logs' );
+			}
+		}
+
+		delete_option( 'el_bundle_license' );
+		$wpdb->query( "DELETE FROM {$wpdb->options} WHERE option_name LIKE 'el_license_%'" );
+	}
 }