+++ HISTORY.md
@@ -1,5 +1,13 @@
 ## Changelog ##
 
+### v2.1.0 - (2017-09-21) ###
+- New: GUI option to choose the user roles that can access email logs.
+- New: GUI option to delete email log table when the plugin is uninstalled.
+- Tweak: Performance improvements.
+- Tweak: Delete all traces of the plugin from DB if the user chooses to destroy data during uninstall.
+- Fix: Handle cases where there is a quote in front of email address.
+- Fix: Handle cases where array passed to `wp_mail` may not contain all the required fields.
+
 ### v2.0.2 - (2017-08-07) ###
 - Fix: Renamed include/util directory to correct case. This caused issues in some install.
 
@@ -146,6 +154,9 @@
 - Initial Release
 
 ## Upgrade Notice ##
+
+### 2.1.0 ###
+GUI option to choose who can access email logs and performance improvements.
 
 ### 2.0.2 ###
 Fixed the case of the Util directory. This caused issues in some install.